Fixed threading
