+++ composeApp/src/iosMain/kotlin/org/escalaralcoiaicomtat/app/ui/platform/MapComposable.ios.kt
@@ -11,18 +11,12 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.interop.UIKitView
-import com.fleeksoft.ksoup.Ksoup
 import io.github.aakira.napier.Napier
 import kotlinx.cinterop.ExperimentalForeignApi
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.IO
-import kotlinx.coroutines.launch
 import org.escalaralcoiaicomtat.app.data.generic.LatLng
 import org.escalaralcoiaicomtat.app.map.placemark.Placemark
 import org.escalaralcoiaicomtat.app.map.style.Style
 import org.escalaralcoiaicomtat.app.map.utils.coordinateRegionOf
-import org.escalaralcoiaicomtat.app.maps.KMZHandler
 import org.escalaralcoiaicomtat.app.ui.reusable.CircularProgressIndicatorBox
 import platform.MapKit.MKAnnotationProtocol
 import platform.MapKit.MKMapTypeSatellite
@@ -34,37 +28,13 @@ data class MapData(
     val styles: List<Style>
 )
 
-private suspend inline fun loadKMZ(
-    kmz: Uuid,
-    onDocumentLoaded: (data: MapData) -> Unit
-) {
-    val kmlFile = KMZHandler.load(kmz, replaceImagePaths = false)
-    val kmlString = kmlFile.read("doc.kml")?.decodeToString()
-    val xml = Ksoup.parse(kmlString)
-    val document = xml.root().getElementsByTag("Document")[0]
-    val styles = document.getElementsByTag("Style")
-        .also { Napier.d { "Processing ${it.size} style elements..." } }
-        .mapNotNull(Style.Companion::parse)
-    val placemarks = document.getElementsByTag("Folder")
-        .also { Napier.d { "Processing ${it.size} folders..." } }
-        .flatMap { folder ->
-            folder.getElementsByTag("Placemark").map { Placemark.Companion.parse(it) }
-        }
-        .filterNotNull()
-    onDocumentLoaded(
-        MapData(placemarks, styles)
-    )
-}
-
 @OptIn(ExperimentalForeignApi::class)
 @Composable
 actual fun MapComposable(viewModel: MapViewModel, modifier: Modifier, kmz: Uuid?) {
     var mapData by remember { mutableStateOf<MapData?>(null) }
 
     LaunchedEffect(kmz) {
-        if (kmz != null && mapData == null) CoroutineScope(Dispatchers.IO).launch {
-            loadKMZ(kmz) { mapData = it }
-        }
+        if (kmz != null && mapData == null) viewModel.loadKMZ(kmz) { mapData = it }
     }
 
     OutlinedCard(modifier) {
